+++ packages/deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-ux/deploy-config-sub-generator
 
+## 0.5.28
+
+### Patch Changes
+
+- 62bb798: set conflicter option
+- Updated dependencies [62bb798]
+    - @sap-ux/abap-deploy-config-sub-generator@0.1.116
+    - @sap-ux/cf-deploy-config-sub-generator@0.2.101
+    - @sap-ux/fiori-generator-shared@0.13.47
+    - @sap-ux/deploy-config-generator-shared@0.1.66
+    - @sap-ux/inquirer-common@0.10.4
+    - @sap-ux/odata-service-inquirer@2.13.5
+
 ## 0.5.27
 
 ### Patch Changes