+++ packages/generator-adp/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sap-ux/generator-adp
 
+## 0.7.29
+
+### Patch Changes
+
+- 62bb798: set conflicter option
+- Updated dependencies [62bb798]
+    - @sap-ux/fiori-generator-shared@0.13.47
+    - @sap-ux/inquirer-common@0.10.4
+    - @sap-ux/adp-tooling@0.18.29
+
 ## 0.7.28
 
 ### Patch Changes