+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 0.1.86
+
+### Patch Changes
+
+- 62bb798: set conflicter option
+- Updated dependencies [62bb798]
+    - @sap-ux/fiori-generator-shared@0.13.47
+    - @sap-ux/ui5-library-reference-inquirer@0.4.86
+
 ## 0.1.85
 
 ### Patch Changes